Move tests associated with judgment checkout & checkin into their own file
Tests for client.get_judgment_checkout_status_message(),
client.get_judgment_checkout_status(), client.checkout_judgment(),
client.checkin_judgment() and associated helper method
client.calculate_seconds_until_midnight()
